backend/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.0
SQLAlchemy==2.0.36
aiosqlite==0.20.0
asyncpg==0.29.0
psycopg2-binary==2.9.9
pydantic==2.9.2
passlib[bcrypt]==1.7.4
PyJWT==2.9.0
email-validator==2.1.1
bcrypt==3.2.2
httpx==0.27.0
python-dotenv==1.0.0
mangum==0.17.0
openai>=1.51.0
